Fix a bug in #from_decimal
I misunderstood the instructions in the paper I’m
using for reference. The decimal doesn’t mutate
on each iteration. The divisor is calculated from
the original number at each step.

This also optimises (and simplifies) the iterator.
If we increment the iteration after using it we
can use it directly when calculating the remainder
rather than creating a temporary variable to hold
`iterator + 1`.